+++ changelog.md
@@ -6,6 +6,7 @@
 
 * added: fix for responsive images loaded by JavaScript from image data attributes
 * fixed: call to undefined function `hash_equals()` on environments with obsolete PHP versions (i.e. < 5.6)
+* fixed: don't run the fixer when a WooCommerce download request is detected
 
 ### 2.6.0, 2018-05-08